Tweak the unmarshal algorithm to be more memory efficient.
This modifies the unmarshal algorithm to allocate fewer Value
structures. It also adds string and value "interning" such that only
one is ever created for the immutable values for each decode. This
will deduplicate a considerable amount of the information in RFC7951
structures as large chunks of data tend to share structural components
and the strings are mostly the same.

Fixes VRVDR-50848