Support for Debian11 / Bullseye
'go tool vet' is not supported in Go1.12 onwards.  Furthermore, they have (un)helpfully changed the 'structtags' flag to 'structtag' so unless we can think of some magic for the rules file that will allow us to detect the Go version, I've had to take out this flag and disable the unit test it was needed for. Can restore when we have fully moved over to Bullseye. I don't want to branch for Bullseye as we don't want to have the branch merge overhead / maintenance to deal with.